Platform dialogs: set the options on construction
Platform dialogs hold a shared pointer to an options structure.
Dialog implementations may access the options in various places
(eg. QGtk3FileDialogHelper::selectFile()) and seem to rely that
the options are not null at this point. Thus, pass the options
pointer right away on construction instead of unnecessarily
delaying to applyOptions(), to avoid potential crashes.

Change-Id: I14b0801eb58a9834867dd302a6610e64adeea0cd
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>