Software Adaptation: Improve clip node handling
This commit solves two issues that were coming up with nested clip
nodes.  The first was that determining the intersection of two or more
nested clip nodes was calculated incorrectly.  This was fixed by making
sure the calculated clip region was in the correct coordinate space.

The second issue was because there was no distinction between an empty
unclipped region, and a completely clipped region.  So a property was
added to specify when a RenderableNode had a clipRegion so that the
empty QRegion could be interpreted correctly.  This makes sure that we
don't end up drawing nodes that were culled.

Change-Id: Id7af092f7fc6a4602b44b585b7dee14084136962
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>