Add support for reading GLSL from files with ShaderEffect
An enabler for using file selectors instead of writing GraphicsInfo-based
conditions in the vertexShader and fragmentShader properties.

[ChangeLog][QtQuick] ShaderEffect with OpenGL now supports reading GLSL
shader sources from local files and from the resource system. Whenever a
fragmentShader or vertexShader property is a valid such URL, the value is
treated as a file specification instead of actual source code.

Change-Id: I9df2a397adc8f82f1eb6b83d27e9e3d9d1a02f46
Reviewed-by: Andy Nichols <andy.nichols@qt.io>