Unify ShaderEffect property setting
rendererInterface() should not require isSceneGraphInitialized() to be
true - the API and language queries like graphicsApi() have no need for the
scenegraph, they only need the plugin to be loaded, i.e. that the QQuickWindow
is constructed.

This is the key to be able to make GraphicsInfo report graphicsApi and
shaderType with the correct values as early as possible - meaning as soon
as the item is associated with a window. The initialization of the
scenegraph (the exact timing of which varies backend to backend) does not
matter here.

The fragment and vertex shader property setting is now unified in the two
ShaderEffect implementations:

- If the component is complete, the shader is attempted to be processed
  right from the setter.

- Otherwise the item will trigger processing once the component is
  complete.

- If there is no window when processing is trigerred, it is deferred via
  polish.

To implement item polish handling we need a new virtual in
QQuickItemPrivate since we cannot intrdouce virtuals into the public
classes.

This way one can write a condition (and later potentially use file
selectors) like this:

fragmentShader: GraphicsInfo.shaderType == GraphicsInfo.GLSL ? "..." : ...

without having to worry about getting an unintended value processed due to
GraphicsInfo not yet reporting an up-to-date value.

parseLog() forces, for GL at least, shader processing to prevent autotests
from breaking.

Change-Id: If55c69d746c29cd07348ddad2d6b0f2b5dd7f3a2
Reviewed-by: Andy Nichols <andy.nichols@qt.io>