Remove DialogButtonBox::flat
The Material Design allows mixing standard flat buttons with non-flat
custom buttons. It is enough that the delegate provides flat standard
buttons by default. It is better to not force all buttons to be flat,
but let the user freely specify whether custom buttons are flat.

This API has not yet been released, so it is still safe to remove.

Change-Id: Iec400553ccefb47f20fb98d64919491d9bb27cbe
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>