D3D12: Improve debug layer support
Filter out the annoying RT clear color warning which we have no control
over if the user called setColor on the QQuickWindow and a non-default
render target is in use.

Start breaking on serious error messages and make it possible to break on
warnings as well. (QT_D3D_DEBUG_BREAK_ON_WARNINGS)

All of this is only relevant when the debug layer is enabled.
(QT_D3D_DEBUG)

Change-Id: Idfc6e3f3165ea50b54938d358543763bc7433c7c
Reviewed-by: Andy Nichols <andy.nichols@qt.io>