localstorage: disable warning about tautological comparison
When building with gcc 6.1.1, there is a warning like this:
plugin.cpp:152:126: error: self-comparison always evaluates to true [-Werror=tautological-compare]
DEFINE_OBJECT_VTABLE(QV4::QQmlSqlDatabaseWrapper);

In the definition of DEFINE_OBJECT_VTABLE, the comparison
&classname::SuperClass::static_vtbl == &Object::static_vtbl ? 0 : &classname::SuperClass::static_vtbl.vTable
will be done at runtime. Maybe the macro could somehow do
that at compile time (or maybe the compiler will optimize
it away after detecting that the result is constant);
but for now, this warning breaks the build.

Task-number: QTBUG-53373
Change-Id: I5f8e429a3f1446d3ef931ff8b0c41fcf652f7a02
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>