add QQuickWindowPrivate::removeGrabber()
The goal is to eventually avoid accessing QQuickWindowPrivate's own
variables inside QQuickItemPrivate, and to use itemForTouchPointId
instead of mouseGrabberItem.

PinchArea relies on some strange behavior: when one touch point is
released, it ungrabs the mouse, but keeps the touch point grab.
This seems somewhat inconsistent, but for now we keep one place
(QQuickItem::ungrabMouse) that does the mouse ungrab only.

Change-Id: I46389475da2ecb157508e092f9f3f0cf923881e3
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>