Add explanation to tst_QQuickMouseArea::moveAndReleaseWithoutPress
The test makes sure that no events are delivered after the initial press
is released. In this context QTRY_* makes little sense, since we are
confirming that the defaults have not changed.
Thus rather spin the event loop 11 times and hope that if we would
accidentally deliver events that we shouldn't, it will be caught in that
time.

Change-Id: Iadeed07a81978aa6679e5513c1fe0aebcaeeb997
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>