QQuickItem: Remove QML_LEAK_CHECK
This seems to have been intended as a poor mans massif, but ultimately is rather
pointless, as it isn't able to tell you *where* the items were allocated, just
that some were still allocated -- and only in debug builds, too.

Just remove it and rely on more conventional tools (debuggers, profilers) for
this sort of thing.

Change-Id: Ib87da49c250166616e90cff92be355021a4622cf
Reviewed-by: Volker Krause <volker.krause@kdab.com>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>