QQuickPointerEvent: store pointCount separately
In 649b309ea9ed7d03aa74565d51edb416c23460d9 it was correct that the
pool did not need to be static, because the whole event is a singleton;
but it allowed the actual m_touchPoints vector to keep growing,
rather than resizing to the current number of points.  When one finger
is pressed, then two, then one is released, pointCount must go from
1 -> 2 -> 1.  And yet we'd like to avoid heap-allocating a fresh
QQuickEventTouchPoint object every time a touchpoint is pressed.  The
choices are then to 1) exploit implementation details of QVector to
allow its count() to be different than its stored size but still
guarantee that every pointer it stores points to a long-lived
QQuickEventTouchPoint (the docs do say "Since Qt 5.6, resize() doesn't
shrink the capacity anymore"), which is what we thought we were doing;
2) store the pointCount in a separate variable, so that the prefix of
m_touchPoints are the valid ones while the remainder is the "pool", or
3) hold a separate reusable pool of touchpoints.  The separate
m_pointCount is cheaper than the pool, so let's go with option 2.

Change-Id: Ic59b5393e9f4b05e4ec27db15d6f06253aa820d5
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>