Minor cleanup in location handling
Store the line and column in CompiledData::Location as unsigned values.
The qmlSourceCoordinate() function(s) already now act as normalizers,
mapping values <= 0 to 0 as indicator for a missing/invalid line/column.
Valid values start at 1 and therefore there is no need to store negative
values in the location structure.

Coincidentally this also fixes a bunch of warnings about conversions from
signed to unsigned.

Change-Id: Ic69ff395d4991989aede695f2e8c58903f1bd2bf
Reviewed-by: Lars Knoll <lars.knoll@qt.io>