qquickitemchangelistener_p.h: add a comment for the sake of clarity
QQuickItemChangeListener::itemGeometryChanged() used to have two QRectF
arguments, which were the new and old geometries. In commit e2c296c the
remaining QRectF argument was changed to a diff of the geometry change.
The meaning of the unnamed argument is not obvious from the method
signature, so add an explanatory comment similar to 83cd7c2.

Change-Id: I62043b3399a86be3490e48c83cedff4f9d630652
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Robin Burchell <robin.burchell@viroteck.net>