StackView: allow choosing which transition to run
This is mostly useful for being able to use either push or pop
transition for replace operations, but there doesn't seem to be
any reason to limit the others. So, now it's able to run any of
the push/pop/replace transitions (or none with "immediate") for
any push/pop/replace operation.

[ChangeLog][StackView] Made it possible to choose the visual
transition type for any operation. This allows using for example
push and pop transitions with replace(), which allows implementing
an "infinite" back and forward navigation pattern while keeping
the amount of instantiated items constant.

Change-Id: I93b87cf854c3d6e2f33ef73ef25ffb3456ec27f4
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>