software: Fix potential infinite loop with ShaderEffect
With the software backend we entered an infinite loop when having
ShaderEffect items. The previous change introducing the usage of polish
requests to defer shader processing was a bit too keen on trying over and
over again: there is exactly one retry attempt we need, not more. If the
second fails too and everything looks initialized and ready, then we know
the backend has no shader effect support so it is futile to try again.

Change-Id: Ibaaefd45e5b26d0be42dbe7f9c080363b7a939bf
Reviewed-by: Andy Nichols <andy.nichols@qt.io>