QQuickEventPoint: be a QObject-child of QQuickPointerEvent
The points are conceptually children, so this will help take care of
memory management.  It's also useful to be able to emit or pass
a QQuickEventPoint without losing the context of the event it came from.
To this end, a pointerEvent() accessor is added, which simply does
the static_cast for you.

Change-Id: I71e57655cf1a0e7d741c4099c7eb9fc3a9a76446
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>