Material: fix ripple wave position on mouse ungrab / button cancel
QQuickMaterialRipple compared the anchor item to QQuickWindow's
mouseGrabberItem to ensure that only mouse-pressed buttons would use the
press point, whereas key-pressed buttons would use the center point.
However, this does not work in case the mouse grab was stolen: when
first pressing an ItemDelegate and then dragging the containing ListView,
the ripple wave would always appear from the center point, not from the
press point as it should. Since QQuickAbstractButton::keyPressEvent()
sets the press point as the center of the control, the mouseGrabberItem
check can be simply removed.

Change-Id: Ic24e5e80c9d849f3fe2684013dd40b004ef55b57
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>