Removing mouse grab must not cancel touch grab
When calling setMouseGrab(nullptr) we would ungrab the touch point that
corresponds to the touchMouseId. This is why the code could not call
setMouseGrab(nullptr) before. Remove the wrong removal of the item from
itemForTouchPointId.
This was not a problem so far because every part of the code that cares
about touch would do special handling and not call this function, but
access QQuickWindow's internals instead.

Change-Id: I032065d5b6db8cd85c78a022a168dbd220440f5f
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>