tst_button: stabilize auto-repeat tests
In a busy CI environment, the wait loop of tryCompare/Verify(),
which is based on timers, may slip through more auto-repeats than
expected. Ideally the repeat count should match the expected value
of 2, of course, but we can accept the result as long as there are
at least 2 repeats that verifies that the button actually starts
repeating.

It now throws warnings when it happens, which it has now done several
times in the CI recently. It should basically never happen when running
the tests locally (neither has the test failed when run locally).

Change-Id: I2a7faaca7f11094512d68955b581fdf71b014d92
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>