Touch and mouse handling: don't pre-grab
The touchmouse test needs adjusting since we don't see ungrab events any
more (which makes sense, nobody actively grabbed anything there).

The comment about storing the touchMouseId early enough is invalid - any
item that wants to steal the grab, has to call grab anyway.

The reason to only update the grab when there is no grabber set yet, is
that in the filterChildMouse stuff some other item might steal the
event, setting the grabber there - that must not be overridden when we
come back to the original event handling.

Change-Id: I4f191b16f4d76113d0c88414b5a80eb1baf99be1
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>