Deliver touch points in defined order
[ChangeLog][QtQuick] Touch events are now delivered in a well defined
order: New touch points are delivered first to items under the finger
which was first touching the screen. Then to items that are under the
second finger and finally to common ancestors.
This means that items that are "on top" will get a chance to grab touch
points before any items in the background.

Change-Id: Icf9a163c0183437cdb79040b8513fd746c3a6a44
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>