Fix crash in tst_qqmlextensionplugin on shutdown
On shutdown the test will unload all the plugins it loaded. In the case
of the QtQuick2 plugin we only loaded it but never called registerTypes,
as the test merely sees if the plugin can be instantiated. Consequently
the attempt at unregistering the value type providers will fail because
it was previously never defined.

Note: We can't just let the QQmlValueTypeProvider destructor take care
of the deregistration, as we do not truly unload plugins anymore
(thankfully). However the plugin object will be destroyed and along with
it we should correctly de-register the things we registered earlier,
otherwise when initializing the plugin again, we'll register handers
multiple times.

Task-number: QTBUG-55524
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 23527754d60780ac4830f1acd6a54d3487a2c362)
Change-Id: I52c9e4c33649966c6291fafaa2efc4242ada6788