deprecate QQuickWindow::sendEvent and minimize the internal usages
It seems to be necessary only to handle parent-filtering cases.
Meanwhile, filtering out certain events while sending others
through via QCoreApplication::sendEvent() also seems questionable,
so let's see if anything breaks if we allow all events through.
And I can't think of a use for it in application code, so let's
deprecate it and hope that we can remove it in Qt 6.

[ChangeLog][QtQuick][QQuickWindow] QQuickWindow::sendEvent() is
deprecated.  Use QCoreApplication::sendEvent() directly instead.

Change-Id: Ide8004bcb0003f14cf176d6d7e80a9af8f442813
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>