Move setting lastMousePosition out of QQuickWindowPrivate::translateTouchEvent
This setting of lastMousePosition is highly suspicious and was nicely
hidden. Make it visible, then a check can be added, for example it
shouldn't be set if we have a real mouse at the same time as touch
events, so a check for the right device is certainly in order.

Also remove a redundant .toPoint() which would result in converting to
QPoint which was then cast back to QPointF.

Change-Id: I8333cf366ce623f3beb20f556f31b50904c0112d
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>