V4 Debugger: use engineAdded() instead of engineAboutToBeAdded()
It does not make a difference in functionality, but after engineAdded()
the server won't wait on a mutex anymore. Before this change, if you
managed to send a message to the V4 debugger after the server had
called aboutToBeAdded(), but before it had stopped waiting, you could
produce a deadlock by scheduling an event for the GUI thread that was
never delivered.

This is a cherry-pick of 18c4295e25503ae637a715858de5c94a3d105a92
from 5.7 as apparently the problem also occurs in 5.6.

Change-Id: Ie2343e6da4bd0b8956d41ff8ebd4d7594616ebd1
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>