Fix crashes with attached property access
We currently embed the attached object ids in the generated code, which
is not stable because they come from a hash and the meta-objects of the
attached types are also not part of the check-sum verification. So for
now disable accelerated access in order to proceed with qtquickcontrols2
integrations. This needs to be re-enabled later with proper checking and
stable lookups.

Change-Id: I54a3f0c905ee26fb80f4e2687a2901a11a921b38
Reviewed-by: Lars Knoll <lars.knoll@qt.io>