Restore version 2.0 for the .impl imports
It's a bit of extra hassle, but the type registration can be done so
that we keep the old version available to avoid deliberately breaking
apps that have had to import .impl to gain access to the internal
goodies. For example, someone may have made a copy of one of the
controls as our documentation suggests. :)

Change-Id: I5308d7e74ecebf69da7fe9c6912380f72c3c9a2b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>