Fix popups to appear above QtQuick.Dialogs 1.x
When using a plain QML Window, the Z-value of the popup item
must be higher than the Z-value of the DefaultWindowDecoration
in QtQuick.Dialogs 1.x. When using an ApplicationWindow, the
same applies to the window overlay.

NOTE: The correct Z-value of DefaultWindowDecoration is 1000000,
not 10000! (qtquickcontrols/src/dialogs/qml/DefaultWindowDecoration.qml)

Task-number: QTBUG-55004
Change-Id: Ib5d068ee2c633884f95ba519f77b6ec082220461
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>