texteditor: use a wrapping Flow for the toolbar
A proper toolbar extension would be better, but this is a workaround
to make the window scalable/resizable until we have a better solution
available.

The font button was moved together with the other formatting
buttons, because it looked a bit awkward when it was wrapping
alone on the second row, and in general it caused things looking
somehow mis-aligned when actions were wrapped on multiple rows.

Change-Id: I0c1f432e81d5fde6aa763c13ea9efbb8af68ad50
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>