stabilize tst_TouchMouse::hoverEnabled
The autotest does not generate the mouse movement that will occur
if you touch an actual touchscreen on X11.
QQuickWindowPrivate::flushFrameSynchronousEvents() will hover the
MouseArea as if the mouse had moved, but that doesn't update
QGuiApplicationPrivate::lastCursorPosition.  The test checks that
the synth mouse move in QQuickWindowPrivate::deliverTouchAsMouse()
occurred in response to the touch release, but if
flushFrameSynchronousEvents occurs before these lines
QVERIFY(mouseArea1->hovered());
QVERIFY(!mouseArea2->hovered());
then it will fail because the touchpoint left mouseArea2 hovered.

Also replace QVERIFY with QCOMPARE where we are comparing a value,
so that when it fails we can see why.

Task-number: QTBUG-55350
Change-Id: I6c9c9ff1aa57c199cdc4be59156bd36e487cbd2b
Reviewed-by: Jan Arve Sæther <jan-arve.saether@theqtcompany.com>